Improved examples and the bin2obj extension

+++ examples/06.extensions/05.bin2obj/src/main.cpp
@@ -1,4 +1,3 @@
-
 #include <stdio.h>
 
 
@@ -8,6 +7,6 @@ extern "C" const unsigned long vararrow_down_svg10913588630147233371_size;
 int main(int argc, char** argv) {
 	printf("Hello, World!\n");
 	printf("Resource size is %lu\n", vararrow_down_svg10913588630147233371_size);
-	printf("Resource content is:\n%.*s\n", vararrow_down_svg10913588630147233371_size, vararrow_down_svg10913588630147233371);
+	printf("Resource content is:\n%.*s\n", (int)vararrow_down_svg10913588630147233371_size, vararrow_down_svg10913588630147233371);
 	return 0;
 }
